major tweak list
issues addressed:

  * Atmega Circuit

   1. PWR_EN from sat connector net not connected to any I/O pin on
ATMega

   * Fuel Gauge Circuit

   1. VBATT (BATT) voltage sense trace for the MAX17047 should be a
Kelvin connection - needs to be split and routed separately from the
current carrying portion of trace to a point close to the battery input
   2. CVBATT (C_BM1) should be connected to VBATT's separate kelvin
line and PK-/CSP, not to main current carrying line and GND
   3. CREG (C_BM2) capacitor should be connected to PK-/CSP not GND.
(Figs 5 and 8 data sheet)
   4. THMST_BM should be connected to PK-/CSP- not GND (Fig 5 data
sheet)
   5. PAD should be connected to PK-/CSP not GND

    * JTAG Connector

   1. EAGLE footprint was provided from TC company with incorrect pin
mapping.  Actual numbers should be down-and-up not zig-zag-across.
Need to fix library footprint and re-route traces, for reference, the
TC2050-IDC data sheet is correctly numbered in all views.

    * PCB layout general

   1. All pads in copper pours need thermal cutouts